GameManager, Ending Screen and other changes
If you pick up the coin at the end or you take damage 3 times you get to the ending screen.

Other minor changes: Added 1 coin and reduced the value of the one at the end to 10 points (total of 20 points if you get every coin). Fixed wall slide so the player doesn't get stuck when trying to jump right next to a wall. Tried to show the HP in the HUD, but failed (HealthPointsHUD). Some code improvement and bug fixes